cuda: convert tests to 64-bit M

+++ test/cuda/cufinufft2dmany_test.cu
@@ -18,7 +18,7 @@
 using cufinufft::utils::infnorm;
 
 template <typename T>
-int run_test(int method, int type, int N1, int N2, int ntransf, int maxbatchsize, int M, T tol, T checktol, int iflag) {
+int run_test(int method, int type, int N1, int N2, int ntransf, int maxbatchsize, int64_t M, T tol, T checktol, int iflag) {
     std::cout << std::scientific << std::setprecision(3);
 
     int ier;
@@ -36,12 +36,12 @@ int run_test(int method, int type, int N1, int N2, int ntransf, int maxbatchsize
     auto randm11 = [&eng, &dist11]() { return dist11(eng); };
 
     // Making data
-    for (int i = 0; i < M; i++) {
+    for (int64_t i = 0; i < M; i++) {
         x[i] = M_PI * randm11(); // x in [-pi,pi)
         y[i] = M_PI * randm11();
     }
     if (type == 1) {
-        for (int i = 0; i < ntransf * M; i++) {
+        for (int64_t i = 0; i < ntransf * M; i++) {
             c[i].real(randm11());
             c[i].imag(randm11());
         }
@@ -147,7 +147,7 @@ int run_test(int method, int type, int N1, int N2, int ntransf, int maxbatchsize
         int i = ntransf - 1;                                // // choose some data to check
         int nt1 = (int)(0.37 * N1), nt2 = (int)(0.26 * N2); // choose some mode index to check
         thrust::complex<T> Ft = thrust::complex<T>(0, 0), J = thrust::complex<T>(0.0, iflag);
-        for (int j = 0; j < M; ++j)
+        for (int64_t j = 0; j < M; ++j)
             Ft += c[j + i * M] * exp(J * (nt1 * x[j] + nt2 * y[j])); // crude direct
         int it = N1 / 2 + nt1 + N1 * (N2 / 2 + nt2);                 // index in complex F as 1d array
         rel_error = abs(Ft - fk[it + i * N]) / infnorm(N1, (std::complex<T> *)fk.data() + i * N);
@@ -156,7 +156,7 @@ int run_test(int method, int type, int N1, int N2, int ntransf, int maxbatchsize
         const int t = ntransf - 1;
         thrust::complex<T> *fkstart = fk.data() + t * N1 * N2;
         const thrust::complex<T> *cstart = c.data() + t * M;
-        const int jt = M / 2; // check arbitrary choice of one targ pt
+        const int64_t jt = M / 2; // check arbitrary choice of one targ pt
         const thrust::complex<T> J(0, iflag);
         thrust::complex<T> ct(0, 0);
         int m = 0;
@@ -196,7 +196,7 @@ int main(int argc, char *argv[]) {
     const int N2 = atof(argv[4]);
     const int ntransf = atof(argv[5]);
     const int maxbatchsize = atoi(argv[6]);
-    const int M = atoi(argv[7]);
+    const int64_t M = atoi(argv[7]);
     const double tol = atof(argv[8]);
     const double checktol = atof(argv[9]);
     const char prec = argv[10][0];
